+++ contracts/contracts/RequestManager.sol
@@ -38,7 +38,7 @@ contract RequestManager is Ownable, LpWhitelist, RestrictedCalls {
         uint96 lpFee;
         uint96 protocolFee;
         uint32 activeClaims;
-        uint256 withdrawClaimId;
+        uint96 withdrawClaimId;
         address filler;
         bytes32 fillId;
         mapping(bytes32 => bool) invalidFillIds;
@@ -76,7 +76,7 @@ contract RequestManager is Ownable, LpWhitelist, RestrictedCalls {
         address targetTokenAddress,
         address targetAddress,
         uint256 amount,
-        uint256 nonce,
+        uint96 nonce,
         uint32 validUntil
     );
 
@@ -96,7 +96,7 @@ contract RequestManager is Ownable, LpWhitelist, RestrictedCalls {
     /// .. seealso:: :sol:func:`claimRequest` :sol:func:`challengeClaim`
     event ClaimMade(
         bytes32 indexed requestId,
-        uint256 claimId,
+        uint96 claimId,
         address claimer,
         uint256 claimerStake,
         address lastChallenger,
@@ -112,7 +112,7 @@ contract RequestManager is Ownable, LpWhitelist, RestrictedCalls {
     ///
     /// .. seealso:: :sol:func:`withdraw`
     event ClaimStakeWithdrawn(
-        uint256 claimId,
+        uint96 claimId,
         bytes32 indexed requestId,
         address claimReceiver
     );
@@ -174,7 +174,7 @@ contract RequestManager is Ownable, LpWhitelist, RestrictedCalls {
     uint256 public constant MAX_VALIDITY_PERIOD = 48 hours;
 
     /// withdrawClaimId is set to this value when an expired request gets withdrawn by the sender
-    uint256 public constant CLAIM_ID_WITHDRAWN_EXPIRED = type(uint256).max;
+    uint96 public constant CLAIM_ID_WITHDRAWN_EXPIRED = type(uint96).max;
 
     // Variables
 
@@ -185,7 +185,7 @@ contract RequestManager is Ownable, LpWhitelist, RestrictedCalls {
     /// A counter used to generate request and claim IDs.
     /// The variable holds the most recently used nonce and must
     /// be incremented to get the next nonce
-    uint256 public currentNonce;
+    uint96 public currentNonce;
 
     /// Maps target rollup chain IDs to finality periods.
     /// Finality periods are in seconds.
@@ -195,7 +195,7 @@ contract RequestManager is Ownable, LpWhitelist, RestrictedCalls {
     mapping(bytes32 => Request) public requests;
 
     /// Maps claim IDs to claims.
-    mapping(uint256 => Claim) public claims;
+    mapping(uint96 => Claim) public claims;
 
     /// Maps ERC20 token address to tokens
     mapping(address => Token) public tokens;
@@ -241,7 +241,7 @@ contract RequestManager is Ownable, LpWhitelist, RestrictedCalls {
     }
 
     /// Check whether a given claim ID is valid.
-    modifier validClaimId(uint256 claimId) {
+    modifier validClaimId(uint96 claimId) {
         require(claims[claimId].claimer != address(0), "claimId not valid");
         _;
     }
@@ -316,7 +316,7 @@ contract RequestManager is Ownable, LpWhitelist, RestrictedCalls {
             "Insufficient allowance"
         );
 
-        uint256 nonce = currentNonce + 1;
+        uint96 nonce = currentNonce + 1;
         currentNonce = nonce;
 
         bytes32 requestId = BeamerUtils.createRequestId(
@@ -400,7 +400,7 @@ contract RequestManager is Ownable, LpWhitelist, RestrictedCalls {
         payable
         validRequestId(requestId)
         onlyWhitelist
-        returns (uint256)
+        returns (uint96)
     {
         Request storage request = requests[requestId];
 
@@ -414,7 +414,7 @@ contract RequestManager is Ownable, LpWhitelist, RestrictedCalls {
 
         request.activeClaims += 1;
 
-        uint256 nonce = currentNonce + 1;
+        uint96 nonce = currentNonce + 1;
         currentNonce = nonce;
         uint256 termination = block.timestamp + claimPeriod;
 
@@ -466,7 +466,7 @@ contract RequestManager is Ownable, LpWhitelist, RestrictedCalls {
     /// Alice and Bob, he would have to challenge with a stake of at least 16.
     ///
     /// @param claimId The claim ID.
-    function challengeClaim(uint256 claimId)
+    function challengeClaim(uint96 claimId)
         external
         payable
         validClaimId(claimId)
@@ -541,7 +541,7 @@ contract RequestManager is Ownable, LpWhitelist, RestrictedCalls {
     ///
     /// @param claimId The claim ID.
     /// @return The address of the deposit receiver.
-    function withdraw(uint256 claimId)
+    function withdraw(uint96 claimId)
         external
         validClaimId(claimId)
         returns (address)
@@ -583,14 +583,14 @@ contract RequestManager is Ownable, LpWhitelist, RestrictedCalls {
         return claimReceiver;
     }
 
-    function resolveClaim(uint256 claimId)
+    function resolveClaim(uint96 claimId)
         private
         view
         returns (address, uint256)
     {
         Claim storage claim = claims[claimId];
         Request storage request = requests[claim.requestId];
-        uint256 withdrawClaimId = request.withdrawClaimId;
+        uint96 withdrawClaimId = request.withdrawClaimId;
         address claimer = claim.claimer;
         uint256 claimerStake = claim.claimerStake;
         uint256 challengerStakeTotal = claim.challengerStakeTotal;
@@ -665,7 +665,7 @@ contract RequestManager is Ownable, LpWhitelist, RestrictedCalls {
         return (claimReceiver, ethToTransfer);
     }
 
-    function withdrawDeposit(Request storage request, uint256 claimId) private {
+    function withdrawDeposit(Request storage request, uint96 claimId) private {
         Claim storage claim = claims[claimId];
         address claimer = claim.claimer;
         emit DepositWithdrawn(claim.requestId, claimer);
